requirements.txt
numpy
Pillow
sounddevice
pystray
